deployment/requirements.txt
tensorflow==2.15.1
pillow==10.4.0
numpy==1.26.4